chor:resolve next-auth version conflict implementation
