+++ app/api/auth/[...nextauth]/route.ts
@@ -1,46 +1,6 @@
-import NextAuth from "next-auth";
-import Twitter from "next-auth/providers/twitter";
+// app/api/auth/[...nextauth]/route.ts
+import NextAuth from "next-auth"
+import { authOptions } from "@/_lib/auth"
 
-export const  handler = NextAuth({
-    providers: [
-        Twitter({
-            clientId: process.env.TWITTER_CLIENT_ID!,
-            clientSecret: process.env.TWITTER_CLIENT_SECRET!,
-            version: '2.0',
-        }),
-    ],
-
-    callbacks: {
-        async jwt({ account, token }) {
-            if (account) {
-                token.accessToken === account.access_token,
-                    token.refeshToken === account.refresh_token
-            }
-
-            return token
-        },
-
-        async session({ session }) {
-            return{
-                ...session,
-                user: {
-                    ...session.user,
-                    isActive: true,
-                    isAdmin: true
-                }
-            }
-        },
-
-         async redirect({ url, baseUrl }) {
-            if (url.startsWith("/")) return `${baseUrl}${url}`
-            else if (new URL(url).origin === baseUrl) return url
-            return baseUrl // Default to home page
-        },
-    },
-
-
-    secret: process.env.AUTH_SECRET,
-    debug: process.env.NODE_ENV === 'development',
-})
-
-export { handler as POST, handler as GET }
+const handler = NextAuth(authOptions)
+export { handler as GET, handler as POST }
